sql/schemachanger/scgraph: move edge defitions to a file
Release note: None
